+++ packages/windie-sdk-js/src/runtime/WindieChatSession.ts
@@ -3,3 +3,13 @@
  */
 
 export * from './AgentChatSession.js';
+
+export {
+  AgentChatSession as WindieChatSession,
+} from './AgentChatSession.js';
+
+export type {
+  AgentChatEditInput as WindieChatEditInput,
+  AgentChatRetryInput as WindieChatRetryInput,
+  AgentChatSendInput as WindieChatSendInput,
+} from './AgentChatSession.js';